pyproject.toml
[tool.poetry]
name = "ra-util"
version = "0.0.1"
description = "Various scripts to assist with rAthena development"
authors = ["Eric Liu <latiosworks@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Latiosu/ra-util"
documentation = "https://github.com/Latiosu/ra-util"

[tool.poetry.dependencies]
python = ">=3.6"
click = "^7.1.2"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
